requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
openai==0.28.0
python-dotenv==1.0.1
gtts==2.5.4
requests==2.32.3